+++ CHANGELOG.md
@@ -14,9 +14,9 @@ All notable changes to this project will be documented in this file.
 
 - MapLibre is now the default map mode.
 
-
 ### Fixed
 
+- Map symbol sizes can now be adjusted in MapLibre mode.
 - Added a day/night terminator overlay to MapLibre mode.
 - Added box zoom support to MapLibre mode.
 - Added feature hover tooltips to MapLibre mode.